+++ magda-search-api/src/main/scala/au/csiro/data61/magda/search/SearchQueryer.scala
@@ -1,13 +1,14 @@
 package au.csiro.data61.magda.search
 
 import au.csiro.data61.magda.api.Query
-import au.csiro.data61.magda.api.model.{ RegionSearchResult, SearchResult, OrganisationsSearchResult }
+import au.csiro.data61.magda.api.model.{ RegionSearchResult, SearchResult, OrganisationsSearchResult, AutoCompleteQueryResult }
 import au.csiro.data61.magda.model.misc._
 
 import scala.concurrent.Future
 
 trait SearchQueryer {
   def search(jwtToken: Option[String], query: Query, start: Long, limit: Int, facetSize: Int, tenantId: BigInt): Future[SearchResult]
+  def autoCompleteQuery(jwtToken: Option[String], field: String, inputString: Option[String], size: Option[Int], tenantId: BigInt): Future[AutoCompleteQueryResult]
   def searchFacets(jwtToken: Option[String],facetType: FacetType, facetQuery: Option[String], generalQuery: Query, start: Int, limit: Int, tenantId: BigInt): Future[FacetSearchResult]
   def searchRegions(query: Option[String], regionType: Option[String], lv1Id: Option[String], lv2Id: Option[String], lv3Id: Option[String], lv4Id: Option[String], lv5Id: Option[String], start: Long, limit: Int, tenantId: BigInt): Future[RegionSearchResult]
   def searchOrganisations(query: Option[String], start: Int, limit: Int, tenantId: BigInt): Future[OrganisationsSearchResult]